refactor: move permit/state/atoms.ts to permittableTokensAtom.ts
